need to patch load_services_oldstyle
The problem is that unless you patch load_services_oldstyle, load_settings is
triggered and this overwrites the dictionary injected by load_defaults.  If
load_services_oldstyle is not patched we traverse into the excpet blocks in
load_defaults_and_settings and from there load_settings is triggered. In
load_settings we set self.defaults to {} again, and thereby overwrite the
dictionary injected via patch.