Fix systemv services incorrectly identified as "serverside"
The new could_be_sysv_service() identifies SystemV services that are
not registered with chkconfig. Previously, these would be mistakenly
identified as "serverside". With this approach it is still possible
to have Systemd and SystemV files and
 - have SystemV take precedence by registering with chkconfig
 - have Systemd take precedence by not registering with chkconfig

The implementation requires the file to be executable in order to
prevent false positives like /etc/init.d/functions, which is just a
collection of shell function, not a service. The "service" command
also requires the init scripts to be executable.